Add files via upload
In this exercise, please implement a Deep Neural Network (DNN) model by yourself to recognize Tibetan handwriting numerals. The following table shows the corresponding Arabic numerals with respect to Tibetan numerals. The samples of this dataset can be found in the following gure. The details of this dataset can be referred to: https://github.com/bat67/TibetanMNIST. Please use train.npz as training data and test.npz as test data.